fog stuff
fix no fog issue which was incorrectly passing data into iFog instead of 1 when fog was disabled

testing a move to pixel shader

adjust passthrough template

for move to pixel shader

remove printf

+++ src/core/hle/D3D8/Direct3D9/Direct3D9.cpp
@@ -7541,7 +7541,7 @@ void CxbxUpdateHostVertexShaderConstants()
 	const float fogDensity = XboxRenderStates.GetXboxRenderStateAsFloat(xbox::_X_D3DRENDERSTATETYPE::X_D3DRS_FOGDENSITY);
 	const float fogStart = XboxRenderStates.GetXboxRenderStateAsFloat(xbox::_X_D3DRENDERSTATETYPE::X_D3DRS_FOGSTART);
 	const float fogEnd = XboxRenderStates.GetXboxRenderStateAsFloat(xbox::_X_D3DRENDERSTATETYPE::X_D3DRS_FOGEND);
-	float fogStuff[4] = { (float)fogTableMode, fogDensity, fogStart, fogEnd };
+	float fogStuff[4] = {fogTableMode, fogDensity, fogStart, fogEnd};
 	g_pD3DDevice->SetVertexShaderConstantF(CXBX_D3DVS_CONSTREG_FOGINFO, fogStuff, 1);
 }
